requirements.txt
GitPython==3.1.14
PyYAML==5.4.1
